Move SVGFile to xml package
